refactor: shared/ui 폴더 구조 정리 및 barrel export 제거 (#232)
