+++ omechu-app/src/shared/ui/search/SearchBar.tsx
